requirements.txt
PyYAML
sklearn
torch
torchvision
pytorch-lightning